Various fixes for warnings and minutia
This commit consists of numerous small changes, none significant enough to
merit a commit on their own:

* Added extra braces to quash for-loop variable scoping issues for older
  compilers (or "g++ -fno-for-scope")

* Appeased "unreachable code" warnings, spurious or otherwise

* Added casts to fix integer-variable signedness warnings

* Added a dummy virtual method to the VectorFileWriter class to silence the
  -Wweak-vtables warning from Clang++

* Renamed some parameters in the Expr and GraphicsWindow classes to
  eliminate "parameter shadows a field" warnings

* Removed an inert "0 ||" from a conditional, and changed a "&& 0" into an
  "#if 0"

* Added missing elements to array/struct/class initializers to zap further
  warnings

* Indented some cpp conditionals where appropriate

* Qualified some variables and functions as static to quiet "no previous
  declaration" warnings

* toolbar.cpp needed to #include<icons-proto.h> to fix those same "no
  previous declaration" warnings from icons.h

* Added some casts and const qualifiers to the Win32 code to address
  warnings produced by g++ when compiling under MinGW

* Rewrote Cnf{Freeze,Thaw}Float() to use a union rather than pointer
  aliasing; this makes g++ a lot happier

* Removed redundant #includes from win32/w32util.cpp

* With Jonathan's blessing, shortened the last line of the About dialog
  text to better match the preceding lines

+++ dsc.h
@@ -150,7 +150,7 @@ public:
     void AllocForOneMore(void) {
         if(n >= elemsAllocated) {
             elemsAllocated = (elemsAllocated + 32)*2;
-            elem = (T *)MemRealloc(elem, elemsAllocated*sizeof(elem[0]));
+            elem = (T *)MemRealloc(elem, (size_t)elemsAllocated*sizeof(elem[0]));
         }
     }
 
@@ -249,7 +249,7 @@ public:
     void Add(T *t) {
         if(n >= elemsAllocated) {
             elemsAllocated = (elemsAllocated + 32)*2;
-            elem = (T *)MemRealloc(elem, elemsAllocated*sizeof(elem[0]));
+            elem = (T *)MemRealloc(elem, (size_t)elemsAllocated*sizeof(elem[0]));
         }
 
         int first = 0, last = n;
@@ -268,7 +268,7 @@ public:
         }
         int i = first;
 
-        memmove(elem+i+1, elem+i, (n-i)*sizeof(elem[0]));
+        memmove(elem+i+1, elem+i, (size_t)(n-i)*sizeof(elem[0]));
         elem[i] = *t;
         n++;
     }
